Create FormHelper macro module and move sing_in_user, select_date macro function. Add submit_form carco function. Move Change flash class prefix to alert.
