+++ app/models/invoice.rb
@@ -1,5 +1,5 @@
 class Invoice < ActiveRecord::Base
-  validates :company, presence: true, uniqueness: true
+  validates :company, presence: true
   validates :salesperson, presence: true
   validates :tax, presence: true
   validates :date, presence: true
