+++ app/views/invoices/_form.slim
@@ -8,4 +8,4 @@
     = f.input :date
 
   = link_to I18n.t('button.cancel'), :back, class: 'btn btn-danger'  
-  = f.button :submit, label: I18n.t('controllers.invoices.new.create_invoice_btn'), class: 'btn-primary'
+  = f.button :submit, class: 'btn-primary'
