+++ spec/rails_helper.rb
@@ -16,6 +16,7 @@ RSpec.configure do |config|
   config.include EmailSpec::Helpers
   config.include EmailSpec::Matchers
   config.include Devise::TestHelpers, type: :controller
+  config.include FormHelper, type: :feature
   config.order = "random"
   config.infer_spec_type_from_file_location!
